Remove ImageServiceProvider
