+++ www/src/pages/components/image/react/index.mdx
@@ -11,224 +11,12 @@ import { ComponentHeader, ComponentFooter } from 'components/thumbprint-componen
 ## `Image`
 
 ```jsx
-<ImageServiceProvider id="356206765797793818" aspectRatio="1:1">
-    {({ src, sources, aspectRatio, ref }) => (
-        <Image
-            src={src}
-            sources={sources}
-            aspectRatio={aspectRatio}
-            style={{ border: '10px solid' }}
-            className="b-gray"
-            ref={ref}
-        />
-    )}
-</ImageServiceProvider>
-```
-
-```jsx
-<ImageServiceProvider id="302055784058904755">
-    {({ src, sources, ref }) => (
-        <Image src={src} sources={sources} aspectRatio="780:406" ref={ref} />
-    )}
-</ImageServiceProvider>
-```
-
-```jsx
-<ImageServiceProvider id="302055784058904755">
-    {({ src, sources, ref }) => (
-        <Image src={src} sources={sources} height="200px" objectPosition="top" ref={ref} />
-    )}
-</ImageServiceProvider>
-```
-
-```jsx
-<ImageServiceProvider id="302055784058904755">
-    {({ src, sources, ref }) => (
-        <Image src={src} sources={sources} height="200px" objectFit="contain" ref={ref} />
-    )}
-</ImageServiceProvider>
-```
-
-```jsx
-<ImageServiceProvider id="356301205135450141" format="png">
-    {({ src, sources }) => (
-        <Image
-            src={src}
-            sources={sources}
-            style={{ border: '10px solid' }}
-            aspectRatio="768:432"
-            objectPosition="right"
-            className="b-gray mh6"
-        />
-    )}
-</ImageServiceProvider>
-```
-
-```jsx
-<Image src="https://media.giphy.com/media/3o7qE8co3YAzNUoAW4/giphy.gif" aspectRatio="48:46" />
-```
-
-```jsx
 <Image
-    src="https://i.giphy.com/media/3ohs7KViF6rA4aan5u/giphy.gif"
-    disableLazyLoading
-    aspectRatio="480:354"
+    src="https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/1024.jpeg"
+    alt="Photo of a cat lying down"
+    aspectRatio="823:548"
 />
 ```
-
-<!--
-```jsx
-<ImageServiceProvider id="356301205135450141" format="png">
-    {({ src, sources }) => (
-        <Image
-            src={src}
-            sources={sources}
-            style={{ border: '10px solid' }}
-            aspectRatio="768:432"
-            objectPosition="right"
-            className="b-gray mh6"
-        />
-    )}
-</ImageServiceProvider>
-``` -->
-
-<!--
-<ImageService id="356206765797793818">
-    {({src, srcSet}) => <Avatar src={src} srcSet={srcSet} size="xlarge" />}
-</ImageService>
--->
-
-<!--
-```jsx
-<Image
-    sources={[
-        {
-            type: 'image/webp',
-            srcSet:
-                'https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/120.webp 120w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/320.webp 320w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/400.webp 400w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/640.webp 640w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/768.webp 768w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/1024.webp 1024w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/1366.webp 1366w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/1600.webp 1600w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/1920.webp 1920w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/2200.webp 2200w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/2350.webp 2350w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/2560.webp 2560w',
-        },
-        {
-            type: 'image/jpeg',
-            srcSet:
-                'https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/120.jpeg 120w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/320.jpeg 320w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/400.jpeg 400w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/640.jpeg 640w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/768.jpeg 768w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/1024.jpeg 1024w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/1366.jpeg 1366w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/1600.jpeg 1600w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/1920.jpeg 1920w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/2200.jpeg 2200w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/2350.jpeg 2350w, https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/2560.jpeg 2560w',
-        },
-    ]}
-    aspectRatio="647:431"
-    style={{
-        border: '10px solid #dedede',
-        // height: '200px'
-    }}
-    // src="https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/120.jpeg"
-    // objectFit="cover"
-    // objectPosition="top"
-/>
-``` -->
-
-<!-- ```jsx
-<Image
-    src="https://d1vg1gqh4nkuns.cloudfront.net/i/356206765797793818/width/1920.webp"
-    aspectRatio="627:418"
-    style={{ border: '10px solid' }}
-    className="b-gray"
-/>
-``` -->
-
-<!-- ## `Image`
-
-The `Image` component loads an image from Thumbtack's Image Service. By default, it lazy-loads the image and [uses responsive images](https://developer.mozilla.org/en-US/docs/Learn/HTML/Multimedia_and_embedding/Responsive_images) to load the best size.
-
-### With just an Image Service `id`
-
-```jsx
-<Image id="350357374752325652" />
-```
-
-### With a custom `objectFit`
-
-```jsx
-<Image id="350357374752325652" objectFit={{ height: 200, style: 'cover' }} />
-```
-
-### With a supported `aspectRatio`
-
-```jsx
-<Image id="350357374752325652" aspectRatio="1:1" />
-```
-
-### With lazy-loading turned off
-
-```jsx
-<Image id="356206765797793818" lazyLoad={null} />
-```
-
-### With a `png`
-
-```jsx
-<Image id="356301205135450141" format="png" />
-``` -->
-
-<!-- ### `ResponsiveImage` with Image Service
-
-This example gets an image from Thumbtack's image service and loads it in `jpeg` and `webp`.
-
-```jsx
-<ResponsiveImage srcSet={getImageServiceSrcSet('350357374752325652', ['jpeg', 'webp'])}>
-    {({ ref, sizes, src, srcSet, style }) => (
-        <picture ref={ref} style={style}>
-            {srcSet.map(s => (
-                <source type={s.type} sizes={sizes} srcSet={s.srcSet} key={s.type} />
-            ))}
-            <img src={src} sizes={sizes} alt="" />
-        </picture>
-    )}
-</ResponsiveImage>
-```
-
-### Single image from the Image Service
-
-This example gets a single image from Thumbtack's image service.
-
-```jsx
-<img src={getImageServiceSrc({ id: '350357374752325652', width: 400, format: 'jpeg' })} alt="" />
-```
-
-### Lazy-loaded image
-
-This example lazy-loads an image from `placeimg.com`.
-
-```jsx
-<LazyImage src="https://placeimg.com/640/480/nature" aspectRatio="640:480" alt="Photo of nature">
-    {({ src, alt }) => <img src={src} alt={alt} className="w-100" />}
-</LazyImage>
-```
-
-### Lazy-loaded responsive image
-
-````jsx
-<LazyImage
-    srcSet={getImageServiceSrcSet('350357374752325652', ['jpeg', 'webp'])}
-    aspectRatio="1600:864"
->
-    {({ src: lazyImageSrc, srcSet: lazyImageSrcSet }, startedLoading) => (
-        <ResponsiveImage srcSet={startedLoading ? lazyImageSrcSet : {}}>
-            {({ ref, sizes, src, srcSet, style }) => (
-                <picture>
-                    {srcSet.map(s => (
-                        <source type={s.type} sizes={sizes} srcSet={s.srcSet} key={s.type} />
-                    ))}
-                    <img
-                        src={startedLoading ? src : lazyImageSrc}
-                        sizes={sizes}
-                        alt=""
-                        style={style}
-                        ref={ref}
-                    />
-                </picture>
-            )}
-        </ResponsiveImage>
-    )}
-</LazyImage>
-``` -->
 
 <ComponentFooter data={props.data} />
